Download backups from CDS and language field in code analyzer.
Added a new python githomes script to replace the original shell
script and to pull down backups from the CDS.  There's already
one that's been converted to python, so we need to merge
these and let one of them go.  Also, added support for a programming
language field in the code analyzer and changed its representation
a little bit.  Added a new sql file to clear all the code analyzer
tables, so we no longer need the reset script.